move dispatchServerGameContext
